+++ utils/firestore.js
@@ -156,26 +156,6 @@ export async function getallPatients(hpUserId) {
             }
         });
     }
-    
-    // for each found patient
-    for (const patient of patients) {
-        // retrieve snapshot of their medication collection
-        await userCollection.doc(patient.id).collection("medications").get()
-        .then( snapshot => {
-            let meds = [];
-            snapshot.forEach(documentSnapshot =>{
-                let id = documentSnapshot.id;
-                let data = documentSnapshot.data();
-                meds.push({
-                    'docId' : id,
-                    'medication': data,
-                    'patientId': patient.id,
-                    'isPatient': true,
-                })
-            });
-            patient.medications = meds;
-        });
-    }
     return patients;
     } catch (error) {
         throw error;
